Add commands for journal and flag search
The journal commands are identical to my notes commands, and now
that my rcfile and var repos are synced, it makes sense to have
journal commands on all of my machines.

The flag command will search for flags in the given directory,
and will be useful with both journal and notes for finding
keywords that I've saved in the various files. It will also
be useful with my new happiness tracking method, where I put
keywords in the flag notation so I can see what makes me
happy or sad over time.