Add shortcut to use julia programming language
I installed julia at work to play around with it and
really like the type system and other features I've seen
so I decided to add a command to call it easily that I
can call at home too once I install it there.

Also move prompt to bottom of file so that it is the last
thing set, which means it doesn't get overwritten by the system
rc files.

Also add the timber command and the pull-rebase command. The
pull-rebase command wasn't really meant for my dev desktop, but
this file could be used on a linux local workspace too so I'd
like to keep it around.