Add command for using jshell with brazil
Lets the user use jshell with the brazil build system, by letting
the brazi-path command collect the classpath for all of the jars
used by the project, then using that classpath as an arg to the
jshell command.

What this means is anything you have in your project, or in any
of your projects dependencies, can be initialized in jshell
directly, which lets you interact with your code. This is good
because it means you can manually test your code in a REPL :D